strbuf: fix reset

+++ include/strvbuf.h
@@ -10,7 +10,7 @@ typedef buf_t strvbuf_t;
 strvbuf_t *strvbuf_init(strvbuf_t *buf, uint cap, size_t len, alloc_t alloc);
 void strvbuf_free(strvbuf_t *buf);
 
-void strvbuf_reset(strvbuf_t *buf, uint used);
+void strvbuf_reset(strvbuf_t *buf, size_t used);
 
 int strvbuf_add(strvbuf_t *buf, strv_t strv, size_t *off);
 strv_t strvbuf_get(const strvbuf_t *buf, size_t off);
